integrate influencer detail view

+++ Model/InfluencerDetailModel.swift
@@ -38,8 +38,8 @@ struct PlatformModel: Identifiable{
 struct ProjectModel: Identifiable{
     let id: Int
     let business_photo: String
-    let sum_impressions: Int
-    let sum_reach: Int
+    let sum_impressions: Double
+    let sum_reach: Double
     let businessOwner_photo: String
     let businessOwner_name: String
     let comment: String
